Fix a data race
The State is a pointer, so copying to the new inputs struct resulting it
modification to the state data outside of the lock.

Instead of copying the mutable state, reference it directly from the
Autopilot struct, so that it is obvious that the State is shared.

Also fix the copying of values. Previously Server.Address was being set
on the existing var (which was not a copy). This commit first does a
copy with `state.Server = existing.Server`, then modifies the copy.
Apply the same fix for existingHealth (take the address of the copy).