Ensure that we don’t pretend to have started the background autopilot routine when we didn’t
Previously if autopilot was in the process of shutting down and Start was called we would happily go along as if it was running. The end state after a few more milliseconds of running would be that autopilot is stopped which wasn’t what the caller desired.

In general a caller can wait for the chan returned by Stop to be closed and then they can know a Start should succeed (assuming not also being started by another concurrent go routine)