Fix an issue with the server stabilization time calculations
Previously we tracked the time autopilot was started and if its uptime was less than 110% of the configured stabilization time then we would disregard the stabilization time.

What was missing was the fact that it was possible for autopilot to be started but the first state didn’t get computed until slightly later. This commit modifies the behavior to track the very first time a state was computed and will only use the stabilization time once our first state becomes old enough.