Fix potential bug with state inconsistencies
gathereNextStateInputs would grab the state lock and look at a few things  within the state. Later on when building the next state we had to regain the state lock and look at existing server information.

This commit stores a pointer to the previous state while holding the state lock. This allows for the buildServerState method to become a free function as it no longer needs to go fetch the current state.

Besides that potential bug fix, this commit also contains a few modifications to make testing implementations of the Promoter and ApplicationIntegration interfaces in other repos easier.

Lastly it also contains some documentation on the state computation process and which external dependencies get used. With the docs, authors of other Promoter and ApplicationIntegration interfaces can know what to expect when mocking portions of autopilot for their own testing.